# services/commerce/order-orchestrator-service/api/openapi.yaml
openapi: 3.1.0
info:
  title: Order Orchestrator Service API
  version: 1.0.0
  description: Coordinates accepted order intents into downstream inventory reservations.
servers:
  - url: http://localhost:8083
paths:
  /api/v1/order-workflows/process-pending:
    post:
      operationId: processPendingOrderWorkflows
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Pending order intents processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessPendingWorkflowsResponse'
  /api/v1/order-workflows/{orderIntentId}/process:
    post:
      operationId: processOrderWorkflow
      parameters:
        - in: path
          name: orderIntentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order intent processed into workflow state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderWorkflowResponse'
        '404':
          description: Accepted order intent event not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/order-workflows/{orderIntentId}:
    get:
      operationId: getOrderWorkflow
      parameters:
        - in: path
          name: orderIntentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current orchestration workflow state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderWorkflowResponse'
        '404':
          description: Workflow not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
    ProcessPendingWorkflowsResponse:
      type: object
      required: [processedCount, results]
      properties:
        processedCount:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/OrderWorkflowResponse'
    OrderWorkflowResponse:
      type: object
      required:
        - workflowId
        - orderIntentId
        - sourceOutboxEventId
        - customerId
        - fulfillmentNodeId
        - status
        - acceptedAt
        - createdAt
        - updatedAt
        - reservations
      properties:
        workflowId:
          type: string
          format: uuid
        orderIntentId:
          type: string
          format: uuid
        sourceOutboxEventId:
          type: string
          format: uuid
        customerId:
          type: string
        fulfillmentNodeId:
          type: string
        status:
          type: string
          enum:
            - INVENTORY_RESERVATION_IN_PROGRESS
            - INVENTORY_RESERVED
            - INVENTORY_RESERVATION_FAILED
        failureReason:
          type: string
          nullable: true
        acceptedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/OrderWorkflowReservationResponse'
    OrderWorkflowReservationResponse:
      type: object
      required:
        - workflowReservationId
        - orderIntentItemId
        - inventoryReservationId
        - nodeId
        - sku
        - quantity
        - status
        - createdAt
      properties:
        workflowReservationId:
          type: string
          format: uuid
        orderIntentItemId:
          type: string
          format: uuid
        inventoryReservationId:
          type: string
          format: uuid
        nodeId:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        status:
          type: string
          enum:
            - RESERVED
            - RELEASED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApiError:
      type: object
      required: [code, message, requestId, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: object
          additionalProperties:
            type: string